[#601] views: add views.py files to apps
